src/containers/general-ocr/model-advanced/requirements.txt
requests==2.25.1
six==1.16.0
opencv-python-headless==4.5.3.56
numpy<=1.23.5
onnxruntime-gpu
Pillow==8.4.0
pyclipper==1.3.0
Shapely==1.7.1
base64image==0.5.1
urllib3==1.26.6
python-dateutil==2.8.2
certifi==2022.12.7
idna==2.10
chardet==4.0.0
boto3